refactor: refactor some code as from review
